refactor: format service layer to team standards
