+++ packages/jaree-server-repo/src/main/java/org/jaree/api/company/controller/CompanyController.java
@@ -12,7 +12,7 @@ import lombok.RequiredArgsConstructor;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/v1/company")
+@RequestMapping("/company")
 public class CompanyController {
 
     private final CompanyService companyService;
