+++ packages/jaree-server-repo/src/main/java/org/jaree/api/company/service/CompanyServiceInterface.java
@@ -1,0 +1,9 @@
+package org.jaree.api.company.service;
+
+import org.jaree.api.company.dto.CompanyWithJobOpeningsOutputDTO;
+
+public interface CompanyServiceInterface {
+
+    CompanyWithJobOpeningsOutputDTO getCompanyInfo(String id);
+
+}
