+++ packages/jaree-server-repo/src/main/java/org/jaree/api/jobopening/service/JobOpeningService.java
@@ -1,8 +1,26 @@
 package org.jaree.api.jobopening.service;
 
 import org.jaree.api.jobopening.dto.JobOpeningOutputDTO;
+import org.jaree.api.jobopening.entity.JobOpening;
+import org.jaree.api.jobopening.repository.JobOpeningRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-public interface JobOpeningService {
+import lombok.RequiredArgsConstructor;
 
-    JobOpeningOutputDTO getJobOpeningInfo(String id);
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class JobOpeningService implements JobOpeningServiceInterface {
+
+    private final JobOpeningRepository jobOpeningRepository;
+
+    @Override
+    public JobOpeningOutputDTO getJobOpeningInfo(String id) {
+
+        JobOpening jobOpening = jobOpeningRepository.findByIdWithoutApplications(id)
+            .orElseThrow(() -> new RuntimeException("JobOpening Not Found"));
+
+        return JobOpeningOutputDTO.from(jobOpening);
+    }
 }
