+++ packages/jaree-server-repo/src/main/java/org/jaree/api/jobopening/service/JobOpeningServiceInterface.java
@@ -1,0 +1,8 @@
+package org.jaree.api.jobopening.service;
+
+import org.jaree.api.jobopening.dto.JobOpeningOutputDTO;
+
+public interface JobOpeningServiceInterface {
+
+    JobOpeningOutputDTO getJobOpeningInfo(String id);
+}
